requirements.txt
Django>=1.8,<1.9
django-phonenumber-field==1.1.0
pycountry==1.20
djangorestframework==3.4.6
djangorestframework-gis==0.10.1
django-rest-swagger==2.0.5
